--- a/filters/alpha/rollup.config.js
+++ b/filters/alpha/rollup.config.js
@@ -1,1 +1,1 @@
-export * from '@internal/builder';
+export { default } from '@internal/builder';