--- a/filters/color-matrix/rollup.config.js
+++ b/filters/color-matrix/rollup.config.js
@@ -1,1 +1,1 @@
-export * from '@internal/builder';
+export { default } from '@internal/builder';