--- a/filters/displacement/package.json
+++ b/filters/displacement/package.json
@@ -34,7 +34,6 @@
   },
   "devDependencies": {
     "@internal/builder": "^5.0.0-alpha",
-    "rollup": "^0.50.0",
-    "rollup-watch": "^4.3.1"
+    "rollup": "^0.50.0"
   }
 }