--- a/filters/displacement/rollup.config.js
+++ b/filters/displacement/rollup.config.js
@@ -1,1 +1,1 @@
-export * from '@internal/builder';
+export { default } from '@internal/builder';