--- a/filters/fxaa/package.json
+++ b/filters/fxaa/package.json
@@ -19,8 +19,8 @@
     "access": "public"
   },
   "scripts": {
-    "build:dev": "rollup -c && rollup -c -f cjs",
-    "build": "rollup -cp && rollup -cp -f cjs",
+    "build:dev": "rollup -c",
+    "build": "rollup -cp",
     "watch": "rollup -cw",
     "postversion": "npm run build",
     "test": "exit 0"