--- a/filters/fxaa/package.json
+++ b/filters/fxaa/package.json
@@ -33,7 +33,6 @@
   },
   "devDependencies": {
     "@internal/builder": "^5.0.0-alpha",
-    "rollup": "^0.50.0",
-    "rollup-watch": "^4.3.1"
+    "rollup": "^0.50.0"
   }
 }