--- a/filters/noise/rollup.config.js
+++ b/filters/noise/rollup.config.js
@@ -1,1 +1,1 @@
-export * from '@internal/builder';
+export { default } from '@internal/builder';