--- a/packages/accessibility/package.json
+++ b/packages/accessibility/package.json
@@ -37,7 +37,6 @@
   "devDependencies": {
     "@internal/builder": "^5.0.0-alpha",
     "@internal/tester": "^5.0.0-alpha",
-    "rollup": "^0.50.0",
-    "rollup-watch": "^4.3.1"
+    "rollup": "^0.50.0"
   }
 }