--- a/packages/app/rollup.config.js
+++ b/packages/app/rollup.config.js
@@ -1,1 +1,1 @@
-export * from '@internal/builder';
+export { default } from '@internal/builder';