--- a/packages/core/rollup.config.js
+++ b/packages/core/rollup.config.js
@@ -1,1 +1,1 @@
-export * from '@internal/builder';
+export { default } from '@internal/builder';