--- a/packages/fragments/rollup.config.js
+++ b/packages/fragments/rollup.config.js
@@ -1,1 +1,1 @@
-export * from '@internal/builder';
+export { default } from '@internal/builder';