--- a/packages/interaction/package.json
+++ b/packages/interaction/package.json
@@ -40,7 +40,6 @@
     "@pixi/graphics": "^5.0.0-alpha",
     "@internal/builder": "^5.0.0-alpha",
     "@internal/tester": "^5.0.0-alpha",
-    "rollup": "^0.50.0",
-    "rollup-watch": "^4.3.1"
+    "rollup": "^0.50.0"
   }
 }