--- a/packages/loaders/rollup.config.js
+++ b/packages/loaders/rollup.config.js
@@ -1,1 +1,1 @@
-export * from '@internal/builder';
+export { default } from '@internal/builder';