--- a/packages/mesh/package.json
+++ b/packages/mesh/package.json
@@ -39,7 +39,6 @@
     "@pixi/loaders": "^5.0.0-alpha",
     "@internal/builder": "^5.0.0-alpha",
     "@internal/tester": "^5.0.0-alpha",
-    "rollup": "^0.50.0",
-    "rollup-watch": "^4.3.1"
+    "rollup": "^0.50.0"
   }
 }