--- a/packages/mixin-cache-as-bitmap/rollup.config.js
+++ b/packages/mixin-cache-as-bitmap/rollup.config.js
@@ -1,1 +1,1 @@
-export * from '@internal/builder';
+export { default } from '@internal/builder';