--- a/packages/mixin-get-child-by-name/rollup.config.js
+++ b/packages/mixin-get-child-by-name/rollup.config.js
@@ -1,1 +1,1 @@
-export * from '@internal/builder';
+export { default } from '@internal/builder';