--- a/packages/mixin-get-global-position/rollup.config.js
+++ b/packages/mixin-get-global-position/rollup.config.js
@@ -1,1 +1,1 @@
-export * from '@internal/builder';
+export { default } from '@internal/builder';