--- a/packages/sprite-animated/package.json
+++ b/packages/sprite-animated/package.json
@@ -35,7 +35,6 @@
   },
   "devDependencies": {
     "@internal/builder": "^5.0.0-alpha",
-    "rollup": "^0.50.0",
-    "rollup-watch": "^4.3.1"
+    "rollup": "^0.50.0"
   }
 }