--- a/packages/sprite-animated/rollup.config.js
+++ b/packages/sprite-animated/rollup.config.js
@@ -1,1 +1,1 @@
-export * from '@internal/builder';
+export { default } from '@internal/builder';