--- a/packages/sprite-tiling/rollup.config.js
+++ b/packages/sprite-tiling/rollup.config.js
@@ -1,1 +1,1 @@
-export * from '@internal/builder';
+export { default } from '@internal/builder';