--- a/packages/spritesheet/package.json
+++ b/packages/spritesheet/package.json
@@ -36,7 +36,6 @@
   "devDependencies": {
     "@internal/builder": "^5.0.0-alpha",
     "@internal/tester": "^5.0.0-alpha",
-    "rollup": "^0.50.0",
-    "rollup-watch": "^4.3.1"
+    "rollup": "^0.50.0"
   }
 }