--- a/packages/text-bitmap/rollup.config.js
+++ b/packages/text-bitmap/rollup.config.js
@@ -1,1 +1,1 @@
-export * from '@internal/builder';
+export { default } from '@internal/builder';