--- a/packages/text/rollup.config.js
+++ b/packages/text/rollup.config.js
@@ -1,1 +1,1 @@
-export * from '@internal/builder';
+export { default } from '@internal/builder';