--- a/tools/builder/index.js
+++ b/tools/builder/index.js
@@ -13,23 +13,18 @@ const preprocess = require('rollup-plugin-preprocess').default;
 const pkg = require(path.resolve('./package'));
 const input = 'src/index.js';
 
-const { prod, format, output } = minimist(process.argv.slice(2), {
-    string: ['format', 'output'],
+const { prod, format } = minimist(process.argv.slice(2), {
+    string: ['main-format'],
     boolean: ['prod'],
     default: {
-        format: 'es',
+        'main-format': 'cjs',
         prod: false,
-        output: '',
     },
     alias: {
-        f: 'format',
         p: 'prod',
-        o: 'output',
+        format: 'main-format',
     },
 });
-
-// Allow overriding output, but default to "module" and "main" fields
-const file = output || (format === 'es' ? pkg.module : pkg.main);
 
 const plugins = [
     resolve({
@@ -40,7 +35,7 @@ const plugins = [
     commonjs({
         namedExports: {
             'resource-loader': ['Resource'],
-            'pixi-gl-core': ['GLFramebuffer'],
+            'pixi-gl-core': ['GLFramebuffer'], // TODO: remove pixi-gl-core
         },
     }),
     string({
@@ -104,7 +99,9 @@ if (prod)
 }
 
 const compiled = (new Date()).toUTCString().replace(/GMT/g, 'UTC');
-
+const external = Object.keys(pkg.dependencies || []);
+const sourcemap = true;
+const name = 'PIXI';
 const banner = `/*!
  * ${pkg.name} - v${pkg.version}
  * Compiled ${compiled}
@@ -113,15 +110,28 @@ const banner = `/*!
  * http://www.opensource.org/licenses/mit-license
  */\n`;
 
-module.exports = {
-    banner,
-    name: 'PIXI',
-    input,
-    output: {
-        file,
-        format,
+exports.default = [
+    {
+        banner,
+        name,
+        input,
+        output: {
+            file: pkg.main,
+            format,
+        },
+        external,
+        sourcemap,
+        plugins,
     },
-    external: Object.keys(pkg.dependencies || []),
-    sourcemap: true,
-    plugins,
-};
+    {
+        banner,
+        input,
+        output: {
+            file: pkg.module,
+            format: 'es',
+        },
+        external,
+        sourcemap,
+        plugins,
+    },
+];