Simple change to optimize the build times overall with Rollup this removes double command line call for each build and takes advantage of any internal Rollup caching and uses less system memory.

Also, removes **rollup-watch** which is no longer needed in Rollup 0.46+

## Results

### `npm run build:dev`

Current: 35s
After PR: 27s
**22% faster**

### `npm run build`

Current: 54s
After PR: 44s
**18% faster**