@cursedcoder 2017-11-08
+1, does it affect bundle size btw?