@bigtimebuddy 2017-11-08
It does not affect bundle size, but in general, Rollup produces much smaller builds than the current Browserify approach.